+++ sshd-common/src/main/java/org/apache/sshd/common/config/ConfigFileReaderSupport.java
@@ -78,22 +78,24 @@ public final class ConfigFileReaderSupport {
     // see http://manpages.ubuntu.com/manpages/precise/en/man5/sshd_config.5.html
     public static final String CIPHERS_CONFIG_PROP = "Ciphers";
     public static final String DEFAULT_CIPHERS =
-            "aes128-ctr,aes192-ctr,aes256-ctr,arcfour256,arcfour128,aes128-cbc,3des-cbc,blowfish-cbc,cast128-cbc,aes192-cbc,aes256-cbc,arcfour";
+        "aes128-ctr,aes192-ctr,aes256-ctr,arcfour256,arcfour128,aes128-cbc,3des-cbc"
+            + ",blowfish-cbc,cast128-cbc,aes192-cbc,aes256-cbc,arcfour";
     // see http://manpages.ubuntu.com/manpages/precise/en/man5/sshd_config.5.html
     public static final String MACS_CONFIG_PROP = "MACs";
     public static final String DEFAULT_MACS =
-            "hmac-md5,hmac-sha1,umac-64@openssh.com,hmac-ripemd160,hmac-sha1-96,hmac-md5-96,hmac-sha2-256,hmac-sha2-256-96,hmac-sha2-512,hmac-sha2-512-96";
+        "hmac-md5,hmac-sha1,umac-64@openssh.com,hmac-ripemd160,hmac-sha1-96"
+            + ",hmac-md5-96,hmac-sha2-256,hmac-sha2-256-96,hmac-sha2-512,hmac-sha2-512-96";
     // see http://manpages.ubuntu.com/manpages/precise/en/man5/sshd_config.5.html
     public static final String KEX_ALGORITHMS_CONFIG_PROP = "KexAlgorithms";
     public static final String DEFAULT_KEX_ALGORITHMS =
-            "ecdh-sha2-nistp256,ecdh-sha2-nistp384,ecdh-sha2-nistp521"
-                    + "," + "diffie-hellman-group-exchange-sha256,diffie-hellman-group-exchange-sha1"
-                    // RFC-8268 groups
-                    + "," + "diffie-hellman-group18-sha512,diffie-hellman-group17-sha512"
-                    + "," + "diffie-hellman-group16-sha512,diffie-hellman-group15-sha512"
-                    + "," + "diffie-hellman-group14-sha256"
-                    // Legacy groups
-                    + "," + "diffie-hellman-group14-sha1,diffie-hellman-group1-sha1";
+        "ecdh-sha2-nistp256,ecdh-sha2-nistp384,ecdh-sha2-nistp521"
+            + "," + "diffie-hellman-group-exchange-sha256,diffie-hellman-group-exchange-sha1"
+            // RFC-8268 groups
+            + "," + "diffie-hellman-group18-sha512,diffie-hellman-group17-sha512"
+            + "," + "diffie-hellman-group16-sha512,diffie-hellman-group15-sha512"
+            + "," + "diffie-hellman-group14-sha256"
+            // Legacy groups
+            + "," + "diffie-hellman-group14-sha1,diffie-hellman-group1-sha1";
     // see http://linux.die.net/man/5/ssh_config
     public static final String HOST_KEY_ALGORITHMS_CONFIG_PROP = "HostKeyAlgorithms";
     // see https://tools.ietf.org/html/rfc5656
@@ -124,7 +126,8 @@ public final class ConfigFileReaderSupport {
     }
 
     public static Properties readConfigFile(InputStream input, boolean okToClose) throws IOException {
-        try (Reader reader = new InputStreamReader(NoCloseInputStream.resolveInputStream(input, okToClose), StandardCharsets.UTF_8)) {
+        try (Reader reader = new InputStreamReader(
+                NoCloseInputStream.resolveInputStream(input, okToClose), StandardCharsets.UTF_8)) {
             return readConfigFile(reader, true);
         }
     }