+++ sshd-core/src/main/java/org/apache/sshd/server/auth/keyboard/InteractiveChallenge.java
@@ -125,6 +125,9 @@ public class InteractiveChallenge implements Cloneable {
 
     @Override
     public String toString() {
-        return getInteractionName() + "[" + getInteractionInstruction() + "](" + getLanguageTag() + "): " + getPrompts();
+        return getInteractionName()
+            + "[" + getInteractionInstruction() + "]"
+            + "(" + getLanguageTag() + ")"
+            + ": " + getPrompts();
     }
 }